backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
ollama==0.3.3
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2
websockets==13.1
